requirements.txt
prefect>=3.0.0rc1
plaid-python==24.0.0
pandas==2.2.2
tabulate==0.9.0
snowflake-sqlalchemy==1.6.1